refactor: PlaySong -> MidiSong
